requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-docx==1.1.0
PyPDF2==3.0.1
openai==1.3.7
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
jinja2==3.1.2
plotly==5.17.0
pandas==2.1.4
numpy==1.25.2
setuptools>=65.0.0
